- hostname resoluytion in place - expanded test cases; cover situation where host resolution fails
git-svn-id: https://svnsrv.desy.de/desy/mtca4u/deviceaccess/branches/rebotHostNameResolution@2029 5f601a5c-88b9-47f3-b851-818f6741d00d

+++ tests/scripts/testRebotBackend.sh
@@ -4,9 +4,29 @@
 SERVER_PID=$!
 sleep .1
 
-# run the test now
+# run the test now; mskrebot uses ip address of the server in the dmap file 
 ../bin/testRebotDevice mskrebot ./dummies.dmap
 TEST_RESULT=$? 
 
-kill $SERVER_PID
-exit $TEST_RESULT
+if [ $TEST_RESULT -ne 0 ] ; then # The above test failed; so do not proceed 
+                                 # further and return 
+ kill $SERVER_PID
+ exit $TEST_RESULT
+fi 
+
+# mskrebot1 uses hostname of the server in the dmap file 
+../bin/testRebotDevice mskrebot1 ./dummies.dmap
+TEST_RESULT=$? 
+
+if [ $TEST_RESULT -ne 0 ] ; then # The above test failed; so do not proceed 
+                                 # further and return 
+ kill $SERVER_PID
+ exit $TEST_RESULT
+fi 
+
+# mskrebot2 is a bad hostname. DNS resolution would fail for this entry 
+../bin/testRebotDevice mskrebot2 ./dummies.dmap
+TEST_RESULT=$? 
+
+ kill $SERVER_PID
+ exit $TEST_RESULT
